requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
requests==2.31.0
